refactor: remove web filters barrel
